requirements.txt
loguru==0.7.2
ollama~=0.1.17
openai==0.28.1
pandas~=2.1.1
PyYAML==6.0.2
Requests==2.32.3
numpy~=1.26.0
scipy~=1.14.1